refactor: consolidate imports from ui components for cleaner code structure
