+++ frontend/src/features/billing/pages/PricingPage.tsx
@@ -5,8 +5,7 @@ import React, { useEffect } from "react";
 import { PricingTable } from "@/components/billing";
 import { PageBackground } from "@/components/layout";
 import Navbar from "@/components/layout/Navbar";
-import { CircleQuestionMarkIcon } from "@/components/ui";
-import IconButton from "@/components/ui/IconButton";
+import { CircleQuestionMarkIcon, IconButton } from "@/components/ui";
 import { useUser } from "@/hooks/auth/useAuthQueries";
 import usePageMetadata from "@/hooks/usePageMetadata";
 import { createCheckoutSession } from "@/utils/apiBilling";
